% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaoptim.R
\name{choice_space}
\alias{choice_space}
\title{Specify optimizatino parameter of character type}
\usage{
choice_space(choices, init = choices[1])
}
\arguments{
\item{choices}{A character vector of choices}

\item{init}{Initial value. Currently not used.}
}
\value{
A \code{ChoiceSpace} object
}
\description{
Specify optimizatino parameter of character type
}
\examples{
choice_space(c("Roughness", "Smooth"))
}
