% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epluspar.R
\docType{package}
\name{epluspar-package}
\alias{epluspar-package}
\alias{epluspar}
\title{The 'epluspar' package.}
\description{
Conduct sensitivity analysis and Bayesian calibration of
EnergyPlus models.
}
\references{
A. Chong and K. Menberg, "Guidelines for the Bayesian calibration of building
energy models", Energy and Buildings, vol. 174, pp. 527–547. DOI:
10.1016/j.enbuild.2018.06.028
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ideas-lab-nus/epluspar}
  \item Report bugs at \url{https://github.com/ideas-lab-nus/epluspar}
}

}
\author{
\strong{Maintainer}: Hongyuan Jia \email{hongyuan.jia@bears-berkeley.sg} (\href{https://orcid.org/0000-0002-0075-8183}{ORCID})

Authors:
\itemize{
  \item Adrian Chong \email{bdgczma@nus.edu.sg} (\href{https://orcid.org/0000-0002-9486-4728}{ORCID})
}

}
