% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epScan.R
\docType{package}
\name{epScan-package}
\alias{epScan-package}
\alias{epScan}
\title{The 'epScan' package.}
\description{
Conduct sensitivity analysis and Bayesian calibration of
EnergyPlus models.
}
\references{
A. Chong and K. Menberg, "Guidelines for the Bayesian calibration of building
energy models", Energy and Buildings, vol. 174, pp. 527–547. DOI:
10.1016/j.enbuild.2018.06.028
}
