% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaoptim.R
\name{float_space}
\alias{float_space}
\title{Specify optimizatino parameter of float type}
\usage{
float_space(min, max, init = mean(c(min, max)))
}
\arguments{
\item{min}{Minimum value}

\item{max}{Maximum value}

\item{init}{Initial value. Currently not used.}
}
\value{
A \code{FloatSpace} object
}
\description{
Specify optimizatino parameter of float type
}
\examples{
float_space(1.0, 5.0)
}
