% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaoptim.R
\name{integer_space}
\alias{integer_space}
\title{Specify optimizatino parameter of integer type}
\usage{
integer_space(integers, init = integers[1])
}
\arguments{
\item{integers}{An integer vector.}

\item{init}{Initial value. Currently not used.}
}
\value{
A \code{IntegerSpace} object.
}
\description{
Specify optimizatino parameter of integer type
}
\examples{
integer_space(1:5)
}
