% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaoptim.R
\name{print.ChoiceSpace}
\alias{print.ChoiceSpace}
\title{Print choice parameter}
\usage{
\method{print}{ChoiceSpace}(x, ...)
}
\arguments{
\item{x}{A \code{ChoiceRange} object}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print choice parameter
}
