% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaoptim.R
\name{print.IntegerRange}
\alias{print.IntegerRange}
\title{Print integer parameter}
\usage{
\method{print}{IntegerRange}(x, ...)
}
\arguments{
\item{x}{An \code{IntegerRange} object}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print integer parameter
}
