% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaoptim.R
\name{printf.IntegerRange}
\alias{printf.IntegerRange}
\title{Print integer parameter}
\usage{
printf.IntegerRange(x, ...)
}
\arguments{
\item{x}{An \code{IntegerRange} object}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print integer parameter
}
