% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaoptim.R
\name{recPCrossover}
\alias{recPCrossover}
\title{General Uniform Crossover}
\usage{
recPCrossover(inds, p = 0.1, ...)
}
\arguments{
\item{inds}{\verb{[list of any]} list of two individuals to perform uniform crossover on}

\item{p}{\verb{[numeric(1)]} per-entry probability to perform crossover.}

\item{...}{further arguments passed on to the method.}
}
\value{
\verb{[list of any]} The mutated individuals.
}
\description{
Crossover recombination operator that crosses over each position iid with
prob. \code{p} and can also be used for non-binary operators.
}
