% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\docType{data}
\name{sensitivity}
\alias{sensitivity}
\alias{Sensitivity}
\title{Conduct Sensitivity Analysis for A EnergyPlus Model}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Sensitivity
}
\description{
Conduct Sensitivity Analysis for A EnergyPlus Model
}
\author{
Hongyuan Jia
}
\keyword{datasets}
