% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaoptim.R
\name{stopOnMaxTime}
\alias{stopOnMaxTime}
\title{Stopping on Maximum Time of evaluations}
\usage{
stopOnMaxTime(max.time = NULL)
}
\arguments{
\item{max.time}{\link{\code{integer(1)}} Time limit in seconds. Default: \code{NULL}.}
}
\value{
An \code{ecr_terminator} object
}
\description{
Stop the EA after a given cutoff time.
}
