% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaoptim.R
\docType{class}
\name{GAOptimJob}
\alias{GAOptimJob}
\title{Conduct Multi-Objective Optimization on An EnergyPlus Model}
\description{
\code{GAOptimJob} class provides a prototype of conducting single- or multi-
objective(s) optimizations on an EnergyPlus model using Genetic Algorithm
}
\details{
The basic workflow is basically:
}
\examples{

## ------------------------------------------------
## Method `GAOptimJob$new`
## ------------------------------------------------

\dontrun{
if (eplusr::is_avail_eplus(8.8)) {
    idf_name <- "1ZoneUncontrolled.idf"
    epw_name <-  "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"

    idf_path <- file.path(eplusr::eplus_config(8.8)$dir, "ExampleFiles", idf_name)
    epw_path <- file.path(eplusr::eplus_config(8.8)$dir, "WeatherData", epw_name)

    # create from local files
    GAOptimJob$new(idf_path, epw_path)

    # create from an Idf and an Epw object
    opt <- GAOptimJob$new(eplusr::read_idf(idf_path), eplusr::read_epw(epw_path))
}
}

}
\author{
Hongyuan Jia
}
\section{Super classes}{
\code{\link[eplusr:EplusGroupJob]{eplusr::EplusGroupJob}} -> \code{\link[eplusr:ParametricJob]{eplusr::ParametricJob}} -> \code{GAOptimJob}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GAOptimJob-new}{\code{GAOptimJob$new()}}
\item \href{#method-GAOptimJob-param}{\code{GAOptimJob$param()}}
\item \href{#method-GAOptimJob-apply_measure}{\code{GAOptimJob$apply_measure()}}
\item \href{#method-GAOptimJob-objective}{\code{GAOptimJob$objective()}}
\item \href{#method-GAOptimJob-recombinator}{\code{GAOptimJob$recombinator()}}
\item \href{#method-GAOptimJob-mutator}{\code{GAOptimJob$mutator()}}
\item \href{#method-GAOptimJob-selector}{\code{GAOptimJob$selector()}}
\item \href{#method-GAOptimJob-terminator}{\code{GAOptimJob$terminator()}}
\item \href{#method-GAOptimJob-validate}{\code{GAOptimJob$validate()}}
\item \href{#method-GAOptimJob-run}{\code{GAOptimJob$run()}}
\item \href{#method-GAOptimJob-best_set}{\code{GAOptimJob$best_set()}}
\item \href{#method-GAOptimJob-pareto_set}{\code{GAOptimJob$pareto_set()}}
\item \href{#method-GAOptimJob-population}{\code{GAOptimJob$population()}}
\item \href{#method-GAOptimJob-print}{\code{GAOptimJob$print()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="errors"><a href='../../eplusr/html/EplusGroupJob.html#method-EplusGroupJob-errors'><code>eplusr::EplusGroupJob$errors()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="kill"><a href='../../eplusr/html/EplusGroupJob.html#method-EplusGroupJob-kill'><code>eplusr::EplusGroupJob$kill()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="list_files"><a href='../../eplusr/html/EplusGroupJob.html#method-EplusGroupJob-list_files'><code>eplusr::EplusGroupJob$list_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="list_table"><a href='../../eplusr/html/EplusGroupJob.html#method-EplusGroupJob-list_table'><code>eplusr::EplusGroupJob$list_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="locate_output"><a href='../../eplusr/html/EplusGroupJob.html#method-EplusGroupJob-locate_output'><code>eplusr::EplusGroupJob$locate_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="output_dir"><a href='../../eplusr/html/EplusGroupJob.html#method-EplusGroupJob-output_dir'><code>eplusr::EplusGroupJob$output_dir()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="read_mdd"><a href='../../eplusr/html/EplusGroupJob.html#method-EplusGroupJob-read_mdd'><code>eplusr::EplusGroupJob$read_mdd()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="read_rdd"><a href='../../eplusr/html/EplusGroupJob.html#method-EplusGroupJob-read_rdd'><code>eplusr::EplusGroupJob$read_rdd()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="read_table"><a href='../../eplusr/html/EplusGroupJob.html#method-EplusGroupJob-read_table'><code>eplusr::EplusGroupJob$read_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="report_data"><a href='../../eplusr/html/EplusGroupJob.html#method-EplusGroupJob-report_data'><code>eplusr::EplusGroupJob$report_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="report_data_dict"><a href='../../eplusr/html/EplusGroupJob.html#method-EplusGroupJob-report_data_dict'><code>eplusr::EplusGroupJob$report_data_dict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="status"><a href='../../eplusr/html/EplusGroupJob.html#method-EplusGroupJob-status'><code>eplusr::EplusGroupJob$status()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="tabular_data"><a href='../../eplusr/html/EplusGroupJob.html#method-EplusGroupJob-tabular_data'><code>eplusr::EplusGroupJob$tabular_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="ParametricJob" data-id="cases"><a href='../../eplusr/html/ParametricJob.html#method-ParametricJob-cases'><code>eplusr::ParametricJob$cases()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="ParametricJob" data-id="models"><a href='../../eplusr/html/ParametricJob.html#method-ParametricJob-models'><code>eplusr::ParametricJob$models()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="ParametricJob" data-id="save"><a href='../../eplusr/html/ParametricJob.html#method-ParametricJob-save'><code>eplusr::ParametricJob$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="ParametricJob" data-id="seed"><a href='../../eplusr/html/ParametricJob.html#method-ParametricJob-seed'><code>eplusr::ParametricJob$seed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="ParametricJob" data-id="version"><a href='../../eplusr/html/ParametricJob.html#method-ParametricJob-version'><code>eplusr::ParametricJob$version()</code></a></span></li>
<li><span class="pkg-link" data-pkg="eplusr" data-topic="ParametricJob" data-id="weather"><a href='../../eplusr/html/ParametricJob.html#method-ParametricJob-weather'><code>eplusr::ParametricJob$weather()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAOptimJob-new"></a>}}
\if{latex}{\out{\hypertarget{method-GAOptimJob-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{GAOptimJob} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$new(idf, epw)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{idf}}{A path to an local EnergyPlus IDF file or an \link[eplusr:Idf]{eplusr::Idf} object.}

\item{\code{epw}}{A path to an local EnergyPlus EPW file or an \link[eplusr:Epw]{eplusr::Epw} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{GAOptimJob} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
if (eplusr::is_avail_eplus(8.8)) {
    idf_name <- "1ZoneUncontrolled.idf"
    epw_name <-  "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"

    idf_path <- file.path(eplusr::eplus_config(8.8)$dir, "ExampleFiles", idf_name)
    epw_path <- file.path(eplusr::eplus_config(8.8)$dir, "WeatherData", epw_name)

    # create from local files
    GAOptimJob$new(idf_path, epw_path)

    # create from an Idf and an Epw object
    opt <- GAOptimJob$new(eplusr::read_idf(idf_path), eplusr::read_epw(epw_path))
}
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAOptimJob-param"></a>}}
\if{latex}{\out{\hypertarget{method-GAOptimJob-param}{}}}
\subsection{Method \code{param()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$param(..., .names = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAOptimJob-apply_measure"></a>}}
\if{latex}{\out{\hypertarget{method-GAOptimJob-apply_measure}{}}}
\subsection{Method \code{apply_measure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$apply_measure(measure, ..., .names = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAOptimJob-objective"></a>}}
\if{latex}{\out{\hypertarget{method-GAOptimJob-objective}{}}}
\subsection{Method \code{objective()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$objective(..., .n = NULL, .dir = "min")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAOptimJob-recombinator"></a>}}
\if{latex}{\out{\hypertarget{method-GAOptimJob-recombinator}{}}}
\subsection{Method \code{recombinator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$recombinator(
  ...,
  .float = setwith(ecr::recSBX, eta = 15, p = 0.7),
  .integer = setwith(recPCrossover, p = 0.7),
  .choice = setwith(recPCrossover, p = 0.7)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAOptimJob-mutator"></a>}}
\if{latex}{\out{\hypertarget{method-GAOptimJob-mutator}{}}}
\subsection{Method \code{mutator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$mutator(
  ...,
  .float = setwith(ecr::mutPolynomial, eta = 25, p = 0.1),
  .integer = mutRandomChoice,
  .choice = mutRandomChoice
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAOptimJob-selector"></a>}}
\if{latex}{\out{\hypertarget{method-GAOptimJob-selector}{}}}
\subsection{Method \code{selector()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$selector(
  parent = ecr::selSimple,
  survival = ecr::selNondom,
  strategy = "plus"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAOptimJob-terminator"></a>}}
\if{latex}{\out{\hypertarget{method-GAOptimJob-terminator}{}}}
\subsection{Method \code{terminator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$terminator(
  fun = NULL,
  name,
  message,
  max_gen = NULL,
  max_eval = NULL,
  max_time = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAOptimJob-validate"></a>}}
\if{latex}{\out{\hypertarget{method-GAOptimJob-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$validate(param = NULL, ddy_only = TRUE, verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAOptimJob-run"></a>}}
\if{latex}{\out{\hypertarget{method-GAOptimJob-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$run(
  mu = 20L,
  p_recomb = 0.7,
  p_mut = 0.1,
  dir = NULL,
  wait = TRUE,
  parallel = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAOptimJob-best_set"></a>}}
\if{latex}{\out{\hypertarget{method-GAOptimJob-best_set}{}}}
\subsection{Method \code{best_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$best_set(unique = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAOptimJob-pareto_set"></a>}}
\if{latex}{\out{\hypertarget{method-GAOptimJob-pareto_set}{}}}
\subsection{Method \code{pareto_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$pareto_set(unique = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAOptimJob-population"></a>}}
\if{latex}{\out{\hypertarget{method-GAOptimJob-population}{}}}
\subsection{Method \code{population()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$population()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GAOptimJob-print"></a>}}
\if{latex}{\out{\hypertarget{method-GAOptimJob-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$print()}\if{html}{\out{</div>}}
}

}
}
