% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaoptim.R
\docType{class}
\name{GAOptimJob}
\alias{GAOptimJob}
\title{Conduct Multi-Objective Optimization on An EnergyPlus Model}
\description{
\code{GAOptimJob} class provides a prototype of conducting single- or multi-
objective(s) optimizations on an EnergyPlus model using Genetic Algorithm
}
\details{
The basic workflow is basically:
}
\examples{

## ------------------------------------------------
## Method `GAOptimJob$new`
## ------------------------------------------------

\dontrun{
if (eplusr::is_avail_eplus(8.8)) {
    idf_name <- "1ZoneUncontrolled.idf"
    epw_name <-  "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"

    idf_path <- file.path(eplusr::eplus_config(8.8)$dir, "ExampleFiles", idf_name)
    epw_path <- file.path(eplusr::eplus_config(8.8)$dir, "WeatherData", epw_name)

    # create from local files
    GAOptimJob$new(idf_path, epw_path)

    # create from an Idf and an Epw object
    opt <- GAOptimJob$new(eplusr::read_idf(idf_path), eplusr::read_epw(epw_path))
}
}

}
\author{
Hongyuan Jia
}
\section{Super classes}{
\code{\link[eplusr:EplusGroupJob]{eplusr::EplusGroupJob}} -> \code{\link[eplusr:ParametricJob]{eplusr::ParametricJob}} -> \code{GAOptimJob}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GAOptimJob$new()}}
\item \href{#method-param}{\code{GAOptimJob$param()}}
\item \href{#method-apply_measure}{\code{GAOptimJob$apply_measure()}}
\item \href{#method-objective}{\code{GAOptimJob$objective()}}
\item \href{#method-recombinator}{\code{GAOptimJob$recombinator()}}
\item \href{#method-mutator}{\code{GAOptimJob$mutator()}}
\item \href{#method-selector}{\code{GAOptimJob$selector()}}
\item \href{#method-terminator}{\code{GAOptimJob$terminator()}}
\item \href{#method-validate}{\code{GAOptimJob$validate()}}
\item \href{#method-run}{\code{GAOptimJob$run()}}
\item \href{#method-pareto_front}{\code{GAOptimJob$pareto_front()}}
\item \href{#method-pareto_set}{\code{GAOptimJob$pareto_set()}}
\item \href{#method-population}{\code{GAOptimJob$population()}}
\item \href{#method-plot}{\code{GAOptimJob$plot()}}
\item \href{#method-print}{\code{GAOptimJob$print()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="errors">}\href{../../eplusr/html/EplusGroupJob.html#method-errors}{\code{eplusr::EplusGroupJob$errors()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="kill">}\href{../../eplusr/html/EplusGroupJob.html#method-kill}{\code{eplusr::EplusGroupJob$kill()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="list_table">}\href{../../eplusr/html/EplusGroupJob.html#method-list_table}{\code{eplusr::EplusGroupJob$list_table()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="locate_output">}\href{../../eplusr/html/EplusGroupJob.html#method-locate_output}{\code{eplusr::EplusGroupJob$locate_output()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="output_dir">}\href{../../eplusr/html/EplusGroupJob.html#method-output_dir}{\code{eplusr::EplusGroupJob$output_dir()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="read_mdd">}\href{../../eplusr/html/EplusGroupJob.html#method-read_mdd}{\code{eplusr::EplusGroupJob$read_mdd()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="read_rdd">}\href{../../eplusr/html/EplusGroupJob.html#method-read_rdd}{\code{eplusr::EplusGroupJob$read_rdd()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="read_table">}\href{../../eplusr/html/EplusGroupJob.html#method-read_table}{\code{eplusr::EplusGroupJob$read_table()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="report_data">}\href{../../eplusr/html/EplusGroupJob.html#method-report_data}{\code{eplusr::EplusGroupJob$report_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="report_data_dict">}\href{../../eplusr/html/EplusGroupJob.html#method-report_data_dict}{\code{eplusr::EplusGroupJob$report_data_dict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="status">}\href{../../eplusr/html/EplusGroupJob.html#method-status}{\code{eplusr::EplusGroupJob$status()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="EplusGroupJob" data-id="tabular_data">}\href{../../eplusr/html/EplusGroupJob.html#method-tabular_data}{\code{eplusr::EplusGroupJob$tabular_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="ParametricJob" data-id="models">}\href{../../eplusr/html/ParametricJob.html#method-models}{\code{eplusr::ParametricJob$models()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="ParametricJob" data-id="save">}\href{../../eplusr/html/ParametricJob.html#method-save}{\code{eplusr::ParametricJob$save()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="ParametricJob" data-id="seed">}\href{../../eplusr/html/ParametricJob.html#method-seed}{\code{eplusr::ParametricJob$seed()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="ParametricJob" data-id="version">}\href{../../eplusr/html/ParametricJob.html#method-version}{\code{eplusr::ParametricJob$version()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="eplusr" data-topic="ParametricJob" data-id="weather">}\href{../../eplusr/html/ParametricJob.html#method-weather}{\code{eplusr::ParametricJob$weather()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a \code{GAOptimJob} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$new(idf, epw)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{idf}}{A path to an local EnergyPlus IDF file or an \link[eplusr:Idf]{eplusr::Idf} object.}

\item{\code{epw}}{A path to an local EnergyPlus EPW file or an \link[eplusr:Epw]{eplusr::Epw} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{GAOptimJob} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
if (eplusr::is_avail_eplus(8.8)) {
    idf_name <- "1ZoneUncontrolled.idf"
    epw_name <-  "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"

    idf_path <- file.path(eplusr::eplus_config(8.8)$dir, "ExampleFiles", idf_name)
    epw_path <- file.path(eplusr::eplus_config(8.8)$dir, "WeatherData", epw_name)

    # create from local files
    GAOptimJob$new(idf_path, epw_path)

    # create from an Idf and an Epw object
    opt <- GAOptimJob$new(eplusr::read_idf(idf_path), eplusr::read_epw(epw_path))
}
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-param"></a>}}
\subsection{Method \code{param()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$param(..., .names = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply_measure"></a>}}
\subsection{Method \code{apply_measure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$apply_measure(measure, ..., .names = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-objective"></a>}}
\subsection{Method \code{objective()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$objective(..., .n = NULL, .dir = "min")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recombinator"></a>}}
\subsection{Method \code{recombinator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$recombinator(
  ...,
  .float = setwith(ecr::recSBX, eta = 15, p = 0.7),
  .integer = setwith(recPCrossover, p = 0.7),
  .choice = setwith(recPCrossover, p = 0.7)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mutator"></a>}}
\subsection{Method \code{mutator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$mutator(
  ...,
  .float = setwith(ecr::mutPolynomial, eta = 25, p = 0.1),
  .integer = mutRandomChoice,
  .choice = mutRandomChoice
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-selector"></a>}}
\subsection{Method \code{selector()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$selector(
  parent = ecr::selSimple,
  survival = ecr::selNondom,
  strategy = "plus"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-terminator"></a>}}
\subsection{Method \code{terminator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$terminator(
  fun = NULL,
  name,
  message,
  max_gen = NULL,
  max_eval = NULL,
  max_time = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$validate(param = NULL, verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$run(
  mu = 20L,
  p_recomb = 0.7,
  p_mut = 0.1,
  dir = NULL,
  wait = TRUE,
  parallel = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pareto_front"></a>}}
\subsection{Method \code{pareto_front()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$pareto_front()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pareto_set"></a>}}
\subsection{Method \code{pareto_set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$pareto_set()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-population"></a>}}
\subsection{Method \code{population()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$population()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$plot(pareto_front = TRUE, pareto_set = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAOptimJob$print()}\if{html}{\out{</div>}}
}

}
}
