% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensi_job}
\alias{sensi_job}
\title{Create a Sensitivity Analysis Job}
\usage{
sensi_job(idf, epw)
}
\arguments{
\item{idf}{A path to an local EnergyPlus IDF file or an \code{Idf} object.}

\item{epw}{A path to an local EnergyPlus EPW file or an \code{Epw} object.}
}
\value{
An \code{SensitivityJob} object.
}
\description{
\code{sensi_job()} takes an IDF and EPW as input, and returns an \code{SensitivityJob}
object for conducting sensitivity analysis on an EnergyPlus model. For more
details, please see \link{SensitivityJob}.
}
\examples{
\dontrun{
if (eplusr::is_avail_eplus(8.8)) {
    idf_name <- "1ZoneUncontrolled.idf"
    epw_name <-  "USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw"

    idf_path <- file.path(eplusr::eplus_config(8.8)$dir, "ExampleFiles", idf_name)
    epw_path <- file.path(eplusr::eplus_config(8.8)$dir, "WeatherData", epw_name)

    # create from local files
    sensi_job(idf_path, epw_path)

    # create from an Idf and an Epw object
    sensi_job(read_idf(idf_path), read_epw(epw_path))
}
}
}
\seealso{
\code{\link[=bayes_job]{bayes_job()}} for creating a Bayesican calibration job.
}
\author{
Hongyuan Jia
}
